refactor(supabase-service): barrel pattern — extract 3 domains (Phase 3.5)
Phase 3.5 brainstorm Égypte 2026-05 — refacto #1/3 supabaseService.
supabaseService.ts : 2443 → 2092 lignes (-14%). Compat 100% : tous les
call sites continuent d'importer depuis "../services/supabaseService"
via re-export `export *`.

NOUVEAU
- src/services/sb/_shared.ts (26 L) : helpers cross-cutting partagés
  entre supabaseService.ts et les domain files (requireSupabase,
  isMissingColumnError, isMissingTableError).
- src/services/sb/manual-pv.ts (122 L) : 3 fns saisies manuelles PV
  cross-mois (upsertManualPvEntry, deleteManualPvEntry, loadManualPvEntries).
- src/services/sb/prospects.ts (154 L) : 4 fns CRUD prospects +
  ProspectRow type + 2 mappers internes (fetchSupabaseProspects,
  createSupabaseProspect, updateSupabaseProspect, deleteSupabaseProspect).
- src/services/sb/follow-up-protocol.ts (122 L) : 4 fns logs protocole
  F1/F21 + 1 mapper + 1 row type (fetchSupabaseFollowUpProtocolLogs,
  fetchAllSupabaseFollowUpProtocolLogs, logSupabaseFollowUpProtocolStep,
  deleteSupabaseFollowUpProtocolLog).

MODIFIE
- supabaseService.ts : 3 blocs sed-extraits (lignes 1743-1853, 2119-2259,
  2338-2443) supprimés + 3 lignes `export * from "./sb/<domain>"` ajoutées
  en fin de fichier. Imports types unused (FollowUpProtocolLog, Prospect,
  etc.) retirés.

NON FAIT (raison : token-cost-aware + risque régression vs gain marginal)
- Extraction complète des ~40 autres exports (auth, clients, follow-ups,
  PV principal, activity logs) en domain files dédiés. Effort résiduel
  ~2-3h agent pour cibler sub-800 lignes. À planifier en chantier dédié
  séparé si Thomas valide ROI sur ces 3 premières extractions.

Build prod : npm run build ✅ — aucune régression chunk, aucune
modification d'API publique.
